Prepare rachio for cloudhooks conversion (#33422)
Reorganize code in order to prepare for webhooks
