+++ homeassistant/components/rachio/const.py
@@ -33,6 +33,11 @@ KEY_USERNAME = "username"
 KEY_ZONE_ID = "zoneId"
 KEY_ZONE_NUMBER = "zoneNumber"
 KEY_ZONES = "zones"
+KEY_CUSTOM_SHADE = "customShade"
+KEY_CUSTOM_CROP = "customCrop"
+
+ATTR_ZONE_TYPE = "type"
+ATTR_ZONE_SHADE = "shade"
 
 # Yes we really do get all these exceptions (hopefully rachiopy switches to requests)
 RACHIO_API_EXCEPTIONS = (
@@ -41,3 +46,11 @@ RACHIO_API_EXCEPTIONS = (
     OSError,
     AssertionError,
 )
+
+STATUS_ONLINE = "ONLINE"
+STATUS_OFFLINE = "OFFLINE"
+
+SIGNAL_RACHIO_UPDATE = DOMAIN + "_update"
+SIGNAL_RACHIO_CONTROLLER_UPDATE = SIGNAL_RACHIO_UPDATE + "_controller"
+SIGNAL_RACHIO_ZONE_UPDATE = SIGNAL_RACHIO_UPDATE + "_zone"
+SIGNAL_RACHIO_SCHEDULE_UPDATE = SIGNAL_RACHIO_UPDATE + "_schedule"
